+++ tools/ports/__init__.py
@@ -13,7 +13,6 @@ import subprocess
 import sys
 from inspect import signature
 from pathlib import Path
-from typing import Dict, Set
 from urllib.request import urlopen
 
 from tools import cache, config, shared, system_libs, utils
@@ -22,7 +21,7 @@ from tools.toolchain_profiler import ToolchainProfiler
 
 ports = []
 
-ports_by_name: Dict[str, object] = {}
+ports_by_name: dict[str, object] = {}
 
 ports_needed = set()
 
@@ -287,7 +286,7 @@ class Ports:
   def get_build_dir():
     return system_libs.get_build_dir()
 
-  name_cache: Set[str] = set()
+  name_cache: set[str] = set()
 
   @staticmethod
   def fetch_port_artifact(name, url, sha512hash=None):